Add Eval module & benchmark suites, channels-last flags and export improvements
Added:
  - New birder.eval module with subcommand-based CLI (python -m birder.eval <command>) for evaluation
    - classification (WebDataset support + configurable --num-workers)
    - adversarial (FGSM, PGD, DeepFool, SimBA)
    - benchmarks: awa2, fishnet, flowers102, bioscan5m, fungiclef, nabirds, newt, plankton, plantdoc, plantnet
  - Eval methods in birder.eval.methods: simpleshot, knn, mlp, ami, linear, svm
  - Evaluation datahub entries in birder.datahub.evaluation: AwA2, FishNet, FungiCLEF2023, NABirds, NeWT, SYKE plankton_IFCB_2022, PlantDoc, PlantNet-300K
  - load_pretrained_model_and_transform convenience loader
  - Detection PT2 export support via torch.export (except Faster R-CNN due to RPN control flow)
  - convert-model: --opset and --simplify for ONNX export control
  - predict: --img-loader to select decoding backend (tv / pil)
  - --channels-last support across predict/predict_detection/eval/benchmark and training CLIs

Changes:
  - Detection ImageList.image_sizes is now a (B, 2) torch.Tensor (was list[tuple[int,int]])
  - Sinkhorn queues (SSL): tensors/pointers/full-state are now registered buffers for state-dict save/load

Fixes:
  - MAE Hiera: fix unpatchify pixel order to match [C, H, W] from get_pixel_label_2d
  - CSWin Transformer PT2 export: fix dynamic batch specialization from int() casting; all classification models exportable

+++ birder/net/rope_deit3.py
@@ -223,7 +223,7 @@ class RoPE_DeiT3(DetectorBackbone, PreTrainEncoder, MaskedTokenOmissionMixin, Ma
                 pt_grid_size=self.pt_grid_size,
             ),
             dim=-1,
-        ).to(self.rope.pos_embed.device)
+        ).to(self.rope.pos_embed.device, dtype=self.rope.pos_embed.dtype)
 
     def detection_features(self, x: torch.Tensor) -> dict[str, torch.Tensor]:
         H, W = x.shape[-2:]
