+++ birder/net/rope_vit.py
@@ -648,7 +648,7 @@ class RoPE_ViT(DetectorBackbone, PreTrainEncoder, MaskedTokenOmissionMixin, Mask
                 pt_grid_size=self.pt_grid_size,
             ),
             dim=-1,
-        ).to(self.rope.pos_embed.device)
+        ).to(self.rope.pos_embed.device, dtype=self.rope.pos_embed.dtype)
 
     def freeze(self, freeze_classifier: bool = True, unfreeze_features: bool = False) -> None:
         for param in self.parameters():
